Fix paths in leapfrog_names.txt to include /uds
LeapfrogAI v0.12.0 included a breaking change which updated the directory structure to comply with the Made for UDS standard. Reference: https://github.com/defenseunicorns/leapfrogai/pull/933

This PR updates the paths in leapfrogai_names.txt accordingly. It also removes the `rag` package, which is no longer relevant, and adds the `supabase` package to the list.

Signed-off-by: Barron Stone <7518212+barronstone@users.noreply.github.com>